rename of components
