+++ src/components/app/App.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Main from '../main/Main';
+import Main from '../main/main';
 import PropTypes from "prop-types";
 
 const App = ({offers}) => {
