+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './components/app/App';
+import App from './components/app/app';
 import {getRandomInteger, arrayRandElement} from './utils';
 import {HOUSING_TYPES} from './utils';
